:bug: fix type mismatch (`uint8` to `uint32`)
uint8 と uint32 を間違えました

+++ models/achievement.go
@@ -1,7 +1,7 @@
 package models
 
 type ClubAchievement struct {
-	AchievementID uint8  `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
+	AchievementID uint32 `gorm:"type:int unsigned;not null;primaryKey;autoIncrement"`
 	UUID          string `gorm:"type:char(36);not null"`
 	Achievement   string `gorm:"type:text;not null"`
 }
